좋아요 카운트 인터페이스 함수를 쿼리메서드로 대체

+++ src/main/java/com/valletta/sns/controller/PostController.java
@@ -69,7 +69,7 @@ public class PostController {
     }
 
     @GetMapping("/{postId}/likes")
-    public Response<Integer> likeCount(@PathVariable("postId") Integer postId, Authentication authentication) {
+    public Response<Long> likeCount(@PathVariable("postId") Integer postId, Authentication authentication) {
 
         return Response.success(postService.likeCount(postId));
     }
